Extract bridge package from helper.
Helper is a general name that may represent anything.
Extracting bridge module from it improves the code stucture a little by
having a smaller, focused package, that does what its name suggest.
