+++ pkg/bridge/bridge_test.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package helper
+package bridge
 
 import (
 	. "github.com/onsi/ginkgo/v2"
